set deployments with shared RWX PVCs to the same SELinux level
By default, k8s will assign a random SELinux level to new pods. This
causes new pods sharing the same RWX PVC to relabel the PVC with a new
SELinux level. When this happens, the old pod loses access to the PVC
contents.

The workaround here is to manually set all deployments that share PVCs
to the same SELinux level.